refactor into dedicated kafka producer/consumer files
