+++ plugin-server/src/kafka/config.ts
@@ -1,0 +1,44 @@
+import { GlobalConfig } from 'node-rdkafka'
+import { hostname } from 'os'
+
+import { KafkaConfig } from '../utils/db/hub'
+
+export const RDKAFKA_LOG_LEVEL_MAPPING = {
+    NOTHING: 0,
+    DEBUG: 7,
+    INFO: 6,
+    WARN: 4,
+    ERROR: 3,
+}
+
+export const createRdConnectionConfigFromEnvVars = (kafkaConfig: KafkaConfig): GlobalConfig => {
+    // We get the config from the environment variables. This method should
+    // convert those vars into connection settings that node-rdkafka can use. We
+    // also set the client.id to the hostname of the machine. This is useful for debugging.
+    const config: GlobalConfig = {
+        'client.id': hostname(),
+        'metadata.broker.list': kafkaConfig.KAFKA_HOSTS,
+        'security.protocol': kafkaConfig.KAFKA_SECURITY_PROTOCOL
+            ? (kafkaConfig.KAFKA_SECURITY_PROTOCOL.toLowerCase() as GlobalConfig['security.protocol'])
+            : 'plaintext',
+        'sasl.mechanisms': kafkaConfig.KAFKA_SASL_MECHANISM,
+        'sasl.username': kafkaConfig.KAFKA_SASL_USER,
+        'sasl.password': kafkaConfig.KAFKA_SASL_PASSWORD,
+        'enable.ssl.certificate.verification': false,
+        log_level: RDKAFKA_LOG_LEVEL_MAPPING[kafkaConfig.KAFKAJS_LOG_LEVEL],
+    }
+
+    if (kafkaConfig.KAFKA_TRUSTED_CERT_B64) {
+        config['ssl.ca.pem'] = Buffer.from(kafkaConfig.KAFKA_TRUSTED_CERT_B64, 'base64').toString()
+    }
+
+    if (kafkaConfig.KAFKA_CLIENT_CERT_B64) {
+        config['ssl.key.pem'] = Buffer.from(kafkaConfig.KAFKA_CLIENT_CERT_B64, 'base64').toString()
+    }
+
+    if (kafkaConfig.KAFKA_CLIENT_CERT_KEY_B64) {
+        config['ssl.certificate.pem'] = Buffer.from(kafkaConfig.KAFKA_CLIENT_CERT_KEY_B64, 'base64').toString()
+    }
+
+    return config
+}
